% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.r
\docType{class}
\name{CMEMS.config-class}
\alias{CMEMS.config-class}
\alias{CMEMS.config}
\title{CMEMS interface configuration parameters}
\description{
A class and generator function to store parameters used in configuring downloads from the CMEMS servers
}
\section{Slots}{

\describe{
\item{\code{python}}{Path to the local python installation}

\item{\code{script}}{path to local script for the motuclient. If NULL, then the client is invoked via use of the
"-m motuclient" argument to python (under the assumption that it is already installed)}

\item{\code{user}}{user name (string)}

\item{\code{pwd}}{user password (string)}

\item{\code{auth.mode}}{the authentication mode: [default: cas]. Use "none" for no authentication, "basic" for basic authentication,
cas for Central Authentication Service}

\item{\code{proxy.server}}{the proxy server (url)}

\item{\code{proxy.user}}{the proxy user (string)}

\item{\code{proxy.pwd}}{the proxy password (string)}

\item{\code{socket.timeout}}{Set a timeout on blocking socket operations (float expressing seconds)}

\item{\code{user.agent}}{Set the identification string (user-agent) for HTTP requests. By default this value is 'Python-urllib/x.x' (where x.x is the version of the python interpreter)}

\item{\code{sync.mode}}{Sets the download mode to synchronous (not recommended)}

\item{\code{block.size}}{The block used to download file (integer expressing bytes)}

\item{\code{motu}}{the Motu server to use (url)}

\item{\code{service.id}}{The service identifier (string)}

\item{\code{product.id}}{The product (data set) to download (string)}

\item{\code{variable}}{The variable(s) to download (vector of strings)}

\item{\code{date.min}}{The min date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS])}

\item{\code{date.max}}{The max date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS ])}

\item{\code{latitude.min}}{The southern limit of the spatial subset (string)}

\item{\code{latitude.max}}{The northern limit of the spatial subset (string)}

\item{\code{longitude.min}}{The western limit of the spatial subset (string)}

\item{\code{longitude.max}}{The eastern limit of the spatial subset (string)}

\item{\code{depth.min}}{The shallow limit of the spatial subset}

\item{\code{depth.max}}{The deep limit of the spatial subset}

\item{\code{out.dir}}{The output directory}

\item{\code{out.name}}{The filename to write to}
}}

\seealso{
Details and documentation on the Motu client, including releases of the software to download, can be
 found on the associated GitHub site, \url{https://github.com/clstoulouse/motu-client-python}
}
