% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{CMEMS.download}
\alias{CMEMS.download}
\alias{CMEMS.download.advanced}
\title{Download from CMEMS}
\usage{
CMEMS.download(x, ROI, date.range, out.path, depth.range = NULL, ...)

CMEMS.download.advanced(x, out.dir = NULL, out.name = NULL,
  date.min = NULL, date.max = NULL, latitude.min = NULL,
  latitude.max = NULL, longitude.min = NULL, longitude.max = NULL,
  depth.min = NULL, depth.max = NULL, debug = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{CMEMS.config}} containing the configuration parameters}

\item{ROI}{An extent object from the raster package specifying the region of interest to subset}

\item{date.range}{The maximum and minimum dates (vector of length two of class "Date")}

\item{depth.range}{The maximum and minimum depths specificed as a vector of length 2 (float in the interval [0 ; 2e31 ] -
does not accept 'Surface' as an argument)}

\item{...}{Arguments to be passed on further to the \code{\link{CMEMS.download.advanced}} function. Overwrites any arguments automatically generated
in \code{\link{CMEMS.download}} function.}

\item{out.dir}{The output dir (string)}

\item{out.name}{The output file name (string)}

\item{date.min}{The min date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS])}

\item{date.max}{The max date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS ])}

\item{latitude.min}{The min latitude (float in the interval [-90 ; 90 ])}

\item{latitude.max}{The max latitude (float in the interval [-90 ; 90 ])}

\item{longitude.min}{The min longitude (float in the interval [-180 ; 180 ])}

\item{longitude.max}{The max longitude (float in the interval [-180 ; 180 ])}

\item{depth.min}{The min depth (float in the interval [0 ; 2e31 ] or string 'Surface')}

\item{depth.max}{The max depth (float in the interval [0 ; 2e31 ] or string 'Surface')}

\item{debug}{Allows debugging of the motu client command - builds the command without running it (logical)}

\item{describe.product}{It allows to have all updated information on a dataset. Output is in XML format}
}
\value{
If debug is TRUE, returns the full command to the motu client, ready to be run (via \code{system()}) or checked manually. If
debug is FALSE (the default), runs the command and returns the error code associated with the script.
}
\description{
Provides R interfaces to the Python-based Motu client developed to provide access to CMEMS data. Two interfaces are
provided - the \code{CMEMS.download} interface is a simple interface taking advantage of R classes, while the
\code{\link{CMEMS.download.advanced}} interface provides the full functionality of the motu client.
CMEMS.download.advanced
}
\seealso{
Details and documentation on the Motu client, including releases of the software to download, can be
 found on the associated GitHub site, \url{https://github.com/clstoulouse/motu-client-python}
}

