% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motu_interface.r
\name{CMEMS}
\alias{CMEMS}
\alias{CMEMS.advanced}
\title{CMEMS Interface}
\usage{
CMEMS(ROI, date.range, out.path, depth.range = NULL, ...)

CMEMS.advanced(out.dir = NULL, out.name = NULL, date.min = NULL,
  date.max = NULL, latitude.min = NULL, latitude.max = NULL,
  longitude.min = NULL, longitude.max = NULL, depth.min = NULL,
  depth.max = NULL, sync.mode = NULL, describe.product = NULL,
  block.size = NULL, socket.timeout = NULL, user.agent = NULL,
  debug = FALSE)
}
\arguments{
\item{ROI}{An extent object from the raster package specifying the region of interest to subset}

\item{date.range}{The maximum and minimum dates (vector of length two of class "Date")}

\item{depth.range}{The maximum and minimum depths specificed as a vector of length 2 (float in the interval [0 ; 2e31 ] -
does not accept 'Surface' as an argument)}

\item{...}{Arguments to be passed on further to the \code{CMEMS.advanced} function. Overwrites any arguments automatically generated
in \code{CMEMS} function.}

\item{out.dir}{The output dir (string)}

\item{out.name}{The output file name (string)}

\item{date.min}{The min date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS])}

\item{date.max}{The max date with optional hour resolution (string following format YYYY-MM-DD [HH:MM:SS ])}

\item{latitude.min}{The min latitude (float in the interval [-90 ; 90 ])}

\item{latitude.max}{The max latitude (float in the interval [-90 ; 90 ])}

\item{longitude.min}{The min longitude (float in the interval [-180 ; 180 ])}

\item{longitude.max}{The max longitude (float in the interval [-180 ; 180 ])}

\item{depth.min}{The min depth (float in the interval [0 ; 2e31 ] or string 'Surface')}

\item{depth.max}{The max depth (float in the interval [0 ; 2e31 ] or string 'Surface')}

\item{sync.mode}{Sets the download mode to synchronous (not recommended)}

\item{describe.product}{It allows to have all updated information on a dataset. Output is in XML format}

\item{block.size}{The block used to download file (integer expressing bytes)}

\item{socket.timeout}{Set a timeout on blocking socket operations (float expressing seconds)}

\item{user.agent}{Set the identification string (user-agent) for HTTP requests. By default this value is 'Python-urllib/x.x' (where x.x is the version of the python interpreter)}

\item{debug}{Allows debugging of the motu client command - builds the command without running it (logical)}
}
\value{
If debug is TRUE, returns the full command to the motu client, ready to be run (via \code{system()}) or checked manually. If
debug is FALSE (the default), runs the command and returns the error code associated with the script.
}
\description{
Provides R interfaces to the Python-based Motu client developed to provide access to CMEMS data. Two interfaces are
provided - the \code{CMEMS} interface is a simple interface taking advantage of R classes, while the \code{CMEMS.advanced}
interface provides full functionality of the motu client.
}
\seealso{
Details and documentation on the Motu client, including releases, can be found on the associated GitHub
site, \url{https://github.com/clstoulouse/motu-client-python}
}

