% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motu_interface.r
\name{configure.motu.client}
\alias{configure.motu.client}
\title{Configure MOTU Client}
\usage{
configure.motu.client(python = "python", script = NULL, user = NULL,
  pwd = NULL, auth.mode = "cas", proxy.server = NULL, proxy.user = NULL,
  proxy.pwd = NULL, socket.timeout = NULL, user.agent = NULL)
}
\arguments{
\item{python}{Path to the local python installation}

\item{script}{Path to the local install of motu - normally this would point to the motu-client.py script}

\item{user}{user name (string)}

\item{pwd}{user password (string)}

\item{auth-mode}{the authentication mode: [default: cas]. Use "none" for no authentication, "basic" for basic authentication,
cas for Central Authentication Service}

\item{proxy-server}{the proxy server (url)}

\item{proxy-user}{the proxy user (string)}

\item{proxy-pwd}{the proxy password (string)}

\item{socket-timeout}{Set a timeout on blocking socket operations (float expressing seconds)}

\item{user-agent}{Set the identification string (user-agent) for HTTP requests. By default this value is 'Python-urllib/x.x' (where x.x is the version of the python interpreter)}
}
\description{
Sets parameters to configure thePython-based Motu client
}
\seealso{
Details and documentation on the Motu client, including releases, can be found on the associated GitHub
site, \url{https://github.com/clstoulouse/motu-client-python}
}

