% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{get.motu.client.version}
\alias{get.motu.client.version}
\title{Get Motu client version}
\usage{
get.motu.client.version(x)
}
\arguments{
\item{x}{An object of class \code{\link{CMEMS.config}} containing the configuration 
parameters. At a minimum, the \code{python} and \code{script} slots need to be populated.}
}
\description{
Returns the version number of the motu client specified by x
}
\examples{
## NOT RUN ## 
cfg <- CMEMS.config(python="python",script="motu-client.py")
get.motu.client.version(cfg)
cfg
}
