% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{motuclient}
\alias{motuclient}
\alias{get.motuclient.version}
\alias{check.motuclient}
\title{Motuclient version numbers}
\usage{
get.motuclient.version(x = "missing")

check.motuclient(x)
}
\arguments{
\item{x}{An object of class \code{\link{CMEMS.config}} containing the configuration 
parameters. At a minimum, the \code{python} and \code{script} slots need to be populated.
If missing, the default \code{\link{CMEMS.config}} object is used.}
}
\value{
\code{get.motuclient.version()} returns the raw version string from the motuclient
}
\description{
Checks that the installed motuclient is working and checks its version number.
}
\details{
\code{check.motuclient()}) first checks that the motuclient can be accessed and run. 
It then checks the version number of this installed version against the minimum requirement 
for this package (currently motuclient 1.8.4) and throws an error accordingly.
}
\examples{
\dontrun{
get.motuclient.version()
}
}
