% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{product.description}
\alias{product.description}
\title{Get product details}
\usage{
product.description(x, variable = "missing", quiet = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{CMEMS.config}} containing the configuration 
parameters. At a minimum, the \code{python}, \code{script}, \code{motu}, \code{service.id}
and \code{product.id} slots need to be populated.}

\item{variable}{Specifies the object to return. Valid values are "times" and 
"depths". All other values, including omission, return an xml object describing the product.}

\item{quiet}{Logical, indicating whether to surpress output}
}
\description{
Requests details from the CMEMS servers about a specific product, including temporal coverage
and depth layers
}
\details{
This function returns the description that is associated with the particular product
id. The depths result is self explanatory, but the times result can be hard to work with, as it
is encoded in ISO 8061 format. See \url{https://github.com/clstoulouse/motu#describe-product} 
for details.
}
\examples{
\dontrun{
#Setup a configuration object, using OSTIA as an example
cfg <- CMEMS.config(motu="http://my.cmems-du.eu/motu-web/Motu",
                    service.id = "SST_GLO_SST_L4_REP_OBSERVATIONS_010_011-TDS",
                    product.id = "METOFFICE-GLO-SST-L4-RAN-OBS-SST",
                    variable = c("analysed_sst","analysis_error"))
#Time description is in ISO 8061 format
product.description(cfg,"times")
}
}
