% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{set.product.from.script}
\alias{set.product.from.script}
\title{Get product details from CMEMS command-line script}
\usage{
set.product.from.script(script)
}
\arguments{
\item{script}{A text string provided by the CMEMS website}
}
\description{
Extracts the available details from a CMEMS script
}
\details{
The CMEMS website has subsetting functionality that can be used as a template generate a download script for use with the motu
client (usually accessed by clicking on the "View Script" button when preparing to download). The function \code{set.product.from.script}
saves the hard work of having to figure out the parameters for use with the \code{RCMEMS} package by parsing this command and
extracting the relevant information - are then added to the local configuration using the \code{set.CMEMS.product} command.
}

