% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motu_interface.r
\name{set.product}
\alias{set.CMEMS.product}
\alias{set.product}
\alias{set.product.from.script}
\title{Specify Product}
\usage{
set.CMEMS.product(motu = NULL, service.id = NULL, product.id = NULL,
  variable = NULL)

set.product.from.script(script)
}
\arguments{
\item{motu}{the motu server to use (url)}

\item{service.id}{The service identifier (string)}

\item{product.id}{The product (data set) to download (string)}

\item{variable}{The variable to download (vector of strings)}

\item{script}{A text string provided by the CMEMS website}
}
\description{
Sets the parameters to specify the product to be downloaded
}
\details{
The CMEMS website has subsetting functionality that can be used as a template generate a download script for use with the motu
client (usually accessed by clicking on the "View Script" button when preparing to download). The function \code{set.product.from.script}
saves the hard work of having to figure out the parameters for use with the \code{RCMEMS} package by parsing this command and
extracting the relevant information - are then added to the local configuration using the \code{set.CMEMS.product} command.
}
\seealso{
Details and documentation on the Motu client, including releases, can be found on the associated GitHub
site, \url{https://github.com/clstoulouse/motu-client-python}
}

