% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.r
\name{update}
\alias{update}
\title{Update a CMEMS.config object}
\usage{
\S4method{update}{CMEMS.config}(object, ...)
}
\arguments{
\item{object}{The \code{CMEMS.config} object to serve as the basis for creating the new object}

\item{...}{Slots in the new object - see documentation of \code{\link{CMEMS.config}} for more details}
}
\description{
Updates a CMEMS.config object with new parameters
}
