% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{parse.CMEMS.script}
\alias{parse.CMEMS.script}
\title{Get details from CMEMS command-line script}
\usage{
parse.CMEMS.script(script, short.parameters = FALSE)
}
\arguments{
\item{script}{A text string provided by the CMEMS website}

\item{short.parameters}{A logical variable indicating whether the script provided by CMEMS
uses short, single-letter, argument names (e.g. -u), or long, expanded, argument names 
(e.g. --user). Earlier versions of CMEMS used the short format - the current iteration uses
the more intuitive long argument names. This argument is therefore included to allow compatability
with older script formats.}
}
\description{
Extracts the available details from a download script provided by CMEMS and uses it
to create an CMEMS.config object
}
\details{
The CMEMS website has subsetting functionality that can be used as a template generate a download script for use with the motu
client (usually accessed by clicking on the "View Script" button when preparing to download). This function
saves the hard work of having to figure out the parameters for use with the \code{RCMEMS} package by parsing this command and
extracting the relevant information.
}
