% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEMS_interface.r
\name{parse.CMEMS.script}
\alias{parse.CMEMS.script}
\title{Get details from CMEMS command-line script}
\usage{
parse.CMEMS.script(script)
}
\arguments{
\item{script}{A text string provided by the CMEMS website}
}
\description{
Extracts the available details from a download script provided by CMEMS and uses it
to create an CMEMS.config object
}
\details{
The CMEMS website has subsetting functionality that can be used as a template generate a download script for use with the motu
client (usually accessed by clicking on the "View Script" button when preparing to download). This function
saves the hard work of having to figure out the parameters for use with the \code{RCMEMS} package by parsing this command and
extracting the relevant information.
}
